merge with master

+++ json/src/com/jetbrains/jsonSchema/impl/fixes/SuggestEnumValuesFix.java
@@ -8,6 +8,9 @@ import com.intellij.codeInspection.BatchQuickFix;
 import com.intellij.codeInspection.CommonProblemDescriptor;
 import com.intellij.codeInspection.LocalQuickFix;
 import com.intellij.codeInspection.ProblemDescriptor;
+import com.intellij.json.JsonBundle;
+import com.intellij.json.psi.JsonElementGenerator;
+import com.intellij.json.psi.JsonProperty;
 import com.intellij.openapi.application.WriteAction;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.editor.ex.EditorEx;
@@ -38,7 +41,7 @@ public class SuggestEnumValuesFix implements LocalQuickFix, BatchQuickFix<Common
   @NotNull
   @Override
   public String getFamilyName() {
-    return "Replace with allowed value";
+    return JsonBundle.message("replace.with.allowed.value");
   }
 
   @Nls(capitalization = Nls.Capitalization.Sentence)
@@ -61,7 +64,14 @@ public class SuggestEnumValuesFix implements LocalQuickFix, BatchQuickFix<Common
       prevPrev = prev.getPrevSibling();
     }
     boolean shouldAddWhitespace = myQuickFixAdapter.fixWhitespaceBefore(initialElement, element);
-    WriteAction.run(() -> element.delete());
+    PsiElement parent = element.getParent();
+    boolean isJsonPropName = parent instanceof JsonProperty && ((JsonProperty)parent).getNameElement() == element;
+    if (isJsonPropName) {
+      WriteAction.run(() -> element.replace(new JsonElementGenerator(project).createStringLiteral("")));
+    }
+    else {
+      WriteAction.run(() -> element.delete());
+    }
     EditorEx editor = EditorUtil.getEditorEx(fileEditor);
     assert editor != null;
     // this is a workaround for buggy formatters such as in YAML - it removes the whitespace after ':' when deleting the value
@@ -73,6 +83,9 @@ public class SuggestEnumValuesFix implements LocalQuickFix, BatchQuickFix<Common
         editor.getCaretModel().moveToOffset(offset + 1);
       });
     }
+    if (isJsonPropName) {
+      editor.getCaretModel().moveToOffset(((JsonProperty)parent).getNameElement().getTextOffset() + 1);
+    }
     CodeCompletionHandlerBase.createHandler(CompletionType.BASIC).invokeCompletion(project, editor);
   }
 
@@ -83,15 +96,16 @@ public class SuggestEnumValuesFix implements LocalQuickFix, BatchQuickFix<Common
 
   @Override
   public void applyFix(@NotNull Project project,
-                       @NotNull CommonProblemDescriptor[] descriptors,
+                       CommonProblemDescriptor @NotNull [] descriptors,
                        @NotNull List<PsiElement> psiElementsToIgnore,
                        @Nullable Runnable refreshViews) {
     Editor editor = FileEditorManager.getInstance(project).getSelectedTextEditor();
     if (editor != null) {
-      HintManager.getInstance().showErrorHint(editor, "Sorry, this fix is not available in batch mode");
+      HintManager.getInstance().showErrorHint(editor, JsonBundle.message("sorry.this.fix.is.not.available.in.batch.mode"));
     }
     else {
-      Messages.showErrorDialog(project, "Sorry, this fix is not available in batch mode", "Not Applicable in Batch Mode");
+      Messages.showErrorDialog(project, JsonBundle.message("sorry.this.fix.is.not.available.in.batch.mode"),
+                               JsonBundle.message("not.applicable.in.batch.mode"));
     }
   }
 }
